Refactoring

+++ CSharp/Day07/Equation.cs
@@ -3,7 +3,7 @@ namespace Day07
 {
     internal class Equation
     {
-        public Int128 Result;
+        public Int64 Result;
         public List<long> Values = new List<long>();    
     }
 }
